우병걸/이미지,포스트 게더세더 int->long

+++ src/main/java/com/anabada/neighbor/used/domain/Image.java
@@ -10,7 +10,7 @@ import lombok.Setter;
 @AllArgsConstructor
 @NoArgsConstructor
 public class Image {
-            private int imgId;//이미지번호
-            private int postId;//게시글번호
+            private long imgId;//이미지번호
+            private long postId;//게시글번호
             private String imgUrl;//이미지url
 }
