+++ src/main/java/com/anabada/neighbor/used/domain/Post.java
@@ -11,12 +11,12 @@ import java.util.Date;
 @NoArgsConstructor
 @ToString
 public class Post {
-    	private int postId;//게시글번호
-        private int memberId;//작성자아이디
+    	private long postId;//게시글번호
+        private long memberId;//작성자아이디
         private String title;//제목
         private String content;//내용
         private String postType;//게시글타입
         private Date postDate;//생성일자
         private Date postUpdate;//수정일자
-        private int postView;//조회수
+        private long postView;//조회수
 }
